Request data from the CDC Wonder API

calls the CDC Wonder API and returns a tidy data frame (tibble)

Usage

getData(
  db = "D66",
  querylist = NULL,
  add = TRUE,
  save = FALSE,
  fn = "query.xml",
  agree = TRUE
)

Arguments

db

Indicate the database, either by label, name, or code. To see the complete list of what's available, use `show_databases()`

querylist

The query list supplied must be a list of lists of names and values. By default it is combined with the default query list for that database. To supply a full list and bypass the default query list, set add to TRUE. See here for instructions on how to create this list.

add

If TRUE (default) querylist is combined with the default query list. Set to FALSE to use querylist as a standalone list of query parameters.

save

If TRUE the query list will be saved, defaults to FALSE

fn

filename for saved query list, defaults to query.xml.

agree

Must be set to TRUE to agree to the CDC terms (default is TRUE). To see the terms for the particular dataset click on that dataset here: https://wonder.cdc.gov. sample

Details

Queries for mortality and births statistics from the National Vital Statistics System cannot limit or group results by any location field, such as Region, Division, State or County, or Urbanization (urbanization categories map to specific geographic counties).

For example, in the D76 online database for Detailed Mortality 1999-2013, the location fields are D76.V9, D76.V10 and D76.V27, and the urbanization fields are D76.V11 and D76.V19. These 'sub-national" data fields cannot be grouped by or limited via the API, although these fields are available in the web application.

See https://wonder.cdc.gov/wonder/help/WONDER-API.html for more information.

Examples

mylist <- list(
  list("Group Results By", "Race"),
  list("Tobacco Use", "1")    # yes
)

mydata <- getData("D66", mylist)
#> No encoding supplied: defaulting to UTF-8.
#> Message from query:
#> Any by-variables picked from {0} need to appear in the order listed, and other by-variables can't come between them.
#> Error in getData("D66", mylist): Internal Server Error (HTTP 500).
mydata
#> Error in eval(expr, envir, enclos): object 'mydata' not found